Refactor helpers
Allow lib/rule-data-snapshot.js to use the helpers functions.

`makeNewDataSnap(path, data)` is now `RuleDataSnapshot.create(path, data)`
