+++ lib/helpers.js
@@ -1,60 +1,36 @@
 'use strict';
 
-var RuleDataSnapshot = require('./rule-data-snapshot');
-
-function trimLeft(path) {
+exports.trimLeft = function(path) {
   path = path || '';
 
   return path.replace(/^\/+/, '');
 }
 
-function trimRight(path) {
+exports.trimRight = function(path) {
   path = path || '';
 
   return path.replace(/\/+$/, '');
 }
 
-function trim(path) {
-  return trimLeft(trimRight(path));
+exports.trim = function(path) {
+  return exports.trimLeft(exports.trimRight(path));
 }
 
 exports.pathMerger = function(root, path) {
-  root = trim(root);
-  path = trimLeft(path);
+  root = exports.trim(root);
+  path = exports.trimLeft(path);
 
   return root + '/' + path;
 }
 
 exports.pathSplitter = function(path) {
-  return trimLeft(path).split('/');
+  return exports.trimLeft(path).split('/');
 };
 
 exports.makeNewDataSnap = function(path, newData) {
+  var RuleDataSnapshot = require('./rule-data-snapshot');
 
-  path = trimLeft(path);
-  newData = RuleDataSnapshot.convert(newData)
+  console.log('makeNewDataSnap is deprecated. Use RuleDataSnapshot.create instead.');
 
-  var result;
-
-  if (path.length === 0) {
-    result = newData;
-  } else {
-
-    var workObject = {};
-    result = workObject;
-
-    exports.pathSplitter(path).forEach(function(pathPart, i, pathParts) {
-
-      if (pathParts.length - i === 1) {
-        workObject[pathPart] = newData;
-      } else {
-        workObject[pathPart] = {};
-      }
-      workObject = workObject[pathPart];
-
-    });
-
-  }
-
-  return new RuleDataSnapshot(result);
+  return RuleDataSnapshot.create(path, newData);
 };
