+++ lib/rule-data-snapshot.js
@@ -2,6 +2,7 @@
 'use strict';
 
 var merge = require('lodash.mergewith');
+var helpers = require('./helpers');
 
 function isEmptyObj (obj) {
   return obj && typeof obj == 'object' && Object.keys(obj).length === 0 && obj.constructor === Object;
@@ -22,6 +23,36 @@ function RuleDataSnapshot(data, path) {
   this._path = path;
 
 }
+
+RuleDataSnapshot.create = function(path, newData) {
+
+  path = helpers.trimLeft(path);
+  newData = RuleDataSnapshot.convert(newData)
+
+  var result;
+
+  if (path.length === 0) {
+    result = newData;
+  } else {
+
+    var workObject = {};
+    result = workObject;
+
+    helpers.pathSplitter(path).forEach(function(pathPart, i, pathParts) {
+
+      if (pathParts.length - i === 1) {
+        workObject[pathPart] = newData;
+      } else {
+        workObject[pathPart] = {};
+      }
+      workObject = workObject[pathPart];
+
+    });
+
+  }
+
+  return new RuleDataSnapshot(result);
+};
 
 RuleDataSnapshot.convert = function(data) {
