+++ lib/ruleset.js
@@ -4,8 +4,7 @@
 var Rule = require('./parser/rule'),
   RuleDataSnapshot = require('./rule-data-snapshot'),
   pathSplitter = require('./helpers').pathSplitter,
-  pathMerger = require('./helpers').pathMerger,
-  makeNewDataSnap = require('./helpers').makeNewDataSnap;
+  pathMerger = require('./helpers').pathMerger;
 
 var validRuleKinds = {
   '.read': true,
@@ -238,7 +237,7 @@ Ruleset.prototype.tryWrite = function(path, root, newData, auth, skipWrite, skip
   // write encompasses both the cascading write rules and the
   // non-cascading validate rules
 
-  var newDataRoot = root.merge(makeNewDataSnap(path, newData));
+  var newDataRoot = root.merge(RuleDataSnapshot.create(path, newData));
 
   var result = {
     path: path,
@@ -263,7 +262,7 @@ Ruleset.prototype.tryPatch = function(path, root, newData, auth, skipWrite, skip
   Object.keys(newData).forEach(function(endPath){
     var pathToNode = pathMerger(path, endPath)
 
-    newDataRoot = newDataRoot.merge(makeNewDataSnap(pathToNode, newData[endPath]));
+    newDataRoot = newDataRoot.merge(RuleDataSnapshot.create(pathToNode, newData[endPath]));
     pathsToTest.push(pathToNode);
   });
