+++ lib/test-helpers.js
@@ -4,8 +4,7 @@
 var root, rules,
   debug = true,
   Ruleset = require('./ruleset'),
-  RuleDataSnapshot = require('./rule-data-snapshot'),
-  makeNewDataSnap = require('./helpers').makeNewDataSnap;
+  RuleDataSnapshot = require('./rule-data-snapshot');
 
 var userDefinitions = exports.userDefinitions = {
 
@@ -167,4 +166,6 @@ exports.getFirebaseRules = function() {
   return rules;
 };
 
-exports.makeNewDataSnap = makeNewDataSnap;
+exports.makeNewDataSnap = function() {
+  return RuleDataSnapshot.create.apply(null, arguments)
+};
